Added code to import some tables from WRDS/CCM
With the wrdsClient() working, some code has been added to update the
local SQLite database with data from WRDS/CCM. At the moment the
COMPMASTER, COMPHEAD and CCMXPF_LNKHIST tables are imported. COMPHIST
proved to be too large and needs a different approach.

The WRDS SAS does not support LIMIT x, y queries, so need to think of
something else. May group by thousands of gvkeys at the same time to be
able to get a smaller result set when downloading COMPHIST.