Refactored table updates
Moved common code into a general function that accepts a vector of table names. Optionally add a filter based on either GVKEYs or PERMNOs.

The local database is now using the same table names as the WRDS cloud (without the CRSP. prefix).

Also added a private utility function to compose a date filter to cover the given period.